Segundo/Ejercicios/Proyecto Intermodular II/008-Proyectos/requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
openai==1.3.0
